+++ src/main/java/ec/edu/espe/arquitectura/escolastico/educacion/model/MatriculaNrc.java
@@ -23,7 +23,7 @@ public class MatriculaNrc implements Serializable {
     @Column(name = "estado", nullable = false, length = 3)
     private String estado;
     @Column(name = "numero", nullable = false)
-    private short numero;
+    private Integer numero;
     @Column(name = "costo", nullable = false, precision = 7, scale = 2)
     private BigDecimal costo;
     @JoinColumns({
@@ -48,7 +48,7 @@ public class MatriculaNrc implements Serializable {
         this.pk = eduMatriculaNrcPK;
     }
 
-    public MatriculaNrc(String codMatricula, int codPersona, short codNrc, int codPeriodo, int codDepartamento, int codMateria) {
+    public MatriculaNrc(String codMatricula, int codPersona, Integer codNrc, int codPeriodo, int codDepartamento, int codMateria) {
         this.pk = new MatriculaNrcPK(codMatricula, codPersona, codNrc, codPeriodo, codDepartamento, codMateria);
     }
 
@@ -68,11 +68,11 @@ public class MatriculaNrc implements Serializable {
         this.estado = estado;
     }
 
-    public short getNumero() {
+    public Integer getNumero() {
         return numero;
     }
 
-    public void setNumero(short numero) {
+    public void setNumero(Integer numero) {
         this.numero = numero;
     }
