+++ src/main/java/ec/edu/espe/arquitectura/escolastico/educacion/model/NrcHorarioPK.java
@@ -8,22 +8,22 @@ import javax.persistence.Embeddable;
 public class NrcHorarioPK implements Serializable {
 
     @Column(name = "cod_nrc", nullable = false)
-    private short codNrc;
+    private Integer codNrc;
     @Column(name = "cod_periodo", nullable = false)
-    private int codPeriodo;
+    private Integer codPeriodo;
     @Column(name = "cod_departamento", nullable = false)
-    private int codDepartamento;
+    private Integer codDepartamento;
     @Column(name = "cod_materia", nullable = false)
-    private int codMateria;
+    private Integer codMateria;
     @Column(name = "cod_aula", nullable = false)
-    private int codAula;
+    private Integer codAula;
     @Column(name = "dia_semana", nullable = false, length = 25)
     private String diaSemana;
 
     public NrcHorarioPK() {
     }
 
-    public NrcHorarioPK(short codNrc, int codPeriodo, int codDepartamento, int codMateria, int codAula, String diaSemana) {
+    public NrcHorarioPK(Integer codNrc, Integer codPeriodo, Integer codDepartamento, Integer codMateria, Integer codAula, String diaSemana) {
         this.codNrc = codNrc;
         this.codPeriodo = codPeriodo;
         this.codDepartamento = codDepartamento;
@@ -32,43 +32,43 @@ public class NrcHorarioPK implements Serializable {
         this.diaSemana = diaSemana;
     }
 
-    public short getCodNrc() {
+    public Integer getCodNrc() {
         return codNrc;
     }
 
-    public void setCodNrc(short codNrc) {
+    public void setCodNrc(Integer codNrc) {
         this.codNrc = codNrc;
     }
 
-    public int getCodPeriodo() {
+    public Integer getCodPeriodo() {
         return codPeriodo;
     }
 
-    public void setCodPeriodo(int codPeriodo) {
+    public void setCodPeriodo(Integer codPeriodo) {
         this.codPeriodo = codPeriodo;
     }
 
-    public int getCodDepartamento() {
+    public Integer getCodDepartamento() {
         return codDepartamento;
     }
 
-    public void setCodDepartamento(int codDepartamento) {
+    public void setCodDepartamento(Integer codDepartamento) {
         this.codDepartamento = codDepartamento;
     }
 
-    public int getCodMateria() {
+    public Integer getCodMateria() {
         return codMateria;
     }
 
-    public void setCodMateria(int codMateria) {
+    public void setCodMateria(Integer codMateria) {
         this.codMateria = codMateria;
     }
 
-    public int getCodAula() {
+    public Integer getCodAula() {
         return codAula;
     }
 
-    public void setCodAula(int codAula) {
+    public void setCodAula(Integer codAula) {
         this.codAula = codAula;
     }
